requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alpaca-py==0.30.1
alpaca-trade-api==3.2.0
alpha_vantage==3.0.0
aniso8601==9.0.1
annotated-types==0.7.0
appdirs==1.4.4
APScheduler==3.10.4
asttokens==2.4.1
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
ccxt==4.4.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
config==0.5.1
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
decorator==5.1.1
deprecation==2.1.0
dnspython==2.6.1
duckdb==1.1.0
email_validator==2.2.0
exchange_calendars==4.5.5
executing==2.1.0
filelock==3.16.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-Security==5.5.2
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
Flask-WTF==1.2.1
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
holidays==0.56
html5lib==1.1
huggingface-hub==0.24.6
ibapi==9.81.1.post1
idna==3.8
ijson==3.3.0
importlib_resources==6.4.5
inflection==0.5.1
iniconfig==2.0.0
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonpickle==3.3.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
korean-lunar-calendar==0.3.1
lumibot==3.7.3
lumiwealth-tradier==0.1.9
lxml==5.3.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.1.0
multitasking==0.0.11
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pandas-datareader==0.10.0
pandas_market_calendars==4.4.1
pandas_ta==0.3.14b0
parso==0.8.4
passlib==1.7.4
peewee==3.17.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.2
plotly==5.24.0
pluggy==1.5.0
polygon-api-client==1.14.2
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycares==4.4.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pyluach==2.2.0
pyparsing==3.1.4
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.1
PyYAML==6.0.1
Quandl==3.7.0
quantstats-lumi==0.3.3
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rpds-py==0.20.0
safetensors==0.4.5
scipy==1.14.1
seaborn==0.13.2
setuptools==74.1.2
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.34
sseclient-py==1.8.0
stack-data==0.6.3
sympy==1.13.2
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.4.0
thetadata==0.9.11
tokenizers==0.19.1
toolz==0.12.1
torch==2.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.20
uuid==1.30
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.4
wget==3.2
wsproto==1.2.0
WTForms==3.1.2
yarl==1.11.1
yfinance==0.2.43
